Fixes the NPE occured when two different transactions have the same messaging handler within same test case.
This fix prevents messaging servers to be closed on endTransaction command if messaging handlers are required
to handle more than one transactions. This behaviour is achived by simply introducing a transactionCount integer
that denotes the total number of different transactions that a messaging handler is reposible to handle within a test case.

In EndTransactionStepProcessorActor, transactionCount is checked if its value is 0, meaning all the transactions are
completed. If so, messaging server that is responsible for the connection is closed.

Fixes #3.